Adding working home and login pages
