+++ app/templates/index.html
@@ -7,7 +7,7 @@
 
 				<ul class="on-right">
 					<li class="active">
-						<a href="index.html">Home</a>
+						<a>Home</a>
 					</li>
 					<li>
 						<a href="login.html">Login</a>
