+++ app/templates/login.html
@@ -1,36 +1,17 @@
-<!DOCTYPE html>
-<html>
+ {%extends "base.html" %} {% block main %}
 
-<head>
-  <meta charset="UTF-8">
-  <title>Yummy Recipes Login Form</title>
-
-  <link href="http://getbootstrap.com/dist/css/bootstrap.min.css" rel="stylesheet">
-  <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css">
-  <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.2.1/jquery.min.js"></script>
-  <script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.min.js"></script>
-  <link href="http://getbootstrap.com/examples/jumbotron-narrow/jumbotron-narrow.css" rel="stylesheet">
-  <link rel="stylesheet" href="static/css/style.css" type="text/css" />
-  <link rel='stylesheet prefetch' href='http://netdna.bootstrapcdn.com/bootstrap/3.0.2/css/bootstrap.min.css'>
-
-  <link rel="stylesheet" href="css/style.css">
-
-
-</head>
 
 <body>
-
-  <body background="static/images/2.jpeg">
-    <div class="wrapper">
-      <form class="form-signin">
-        <h2 class="form-signin-heading">login</h2>
-        <input type="text" class="form-control" name="username" placeholder="Email Address" required="" autofocus="" />
-        <input type="password" class="form-control" name="password" placeholder="Password" required="" />
-        <button class="btn btn-lg btn-primary btn-block" type="submit">Login</button>
-      </form>
-    </div>
+  <div class="wrapper">
+    <form class="form-signin">
+      <h2 class="form-signin-heading">login</h2>
+      <input type="text" class="form-control" name="username" placeholder="Email Address" required="" autofocus="" />
+      <input type="password" class="form-control" name="password" placeholder="Password" required="" />
+      <button class="btn btn-lg btn-primary btn-block" type="submit">Login</button>
+    </form>
+  </div>
 
 
-  </body>
+</body>
 
-</html>
+{% endblock %}
