+++ app/views.py
@@ -1,13 +1,17 @@
-from flask import render_template, url_for
-
 from app import app
+from flask import Flask, render_template, url_for
 
 
-@app.route('/')
-def index():
-    return render_template("index.html")
+@app.route("/")
+def main():
+    return render_template('index.html')
 
 
-@app.route('/about')
-def about():
-    return render_template("about.html")
+@app.route("/signup")
+def signup():
+    return render_template('signup.html')
+
+
+@app.route("/login")
+def login():
+    return render_template('login.html')
